refactor(cli): split package exports resolution tests
